# openapi.yaml
openapi: 3.1.0
info:
  title: msbd — Microsandbox REST host
  version: 0.6.0
  description: >
    A standalone, self-hostable HTTP server that wraps the microsandbox Go SDK
    (cgo + libkrun) and exposes a small REST API for managing fast, local
    microVMs. Run it on a host that has /dev/kvm; clients then treat
    microsandbox as a remote sandbox backend with no cgo on their side.


    Every versioned endpoint lives under `/api/v1`. The unversioned paths
    `/healthz`, `/readyz`, `/metrics`, `/docs` and `/openapi.yaml` sit at the
    root, and the root itself (`/`, `/sandboxes`, `/volumes`, `/images`,
    `/snapshots`, `/settings/*`) is the optional web dashboard, which
    authenticates with its own login session rather than a bearer token.


    All endpoints except `/healthz` and `/readyz` require a bearer token.


    Tokens come from either `--api-key` / `MSBD_API_KEY` (comma-separated for
    rotation) or msbd's own key store, managed with `msbd keys create <name>`
    and from the dashboard. Both kinds are accepted together; a stored key can
    be revoked (`msbd keys revoke`) or given an expiry, after which it returns
    401 like any other invalid token. If no key exists at all the server is
    unauthenticated — intended for local development only.
servers:
  - url: http://localhost:8099
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/sandboxes:
    post:
      tags: [lifecycle]
      summary: Create (boot) a sandbox
      description: >
        Boots a detached microVM from an OCI image. Cold first pull of a tag can
        take minutes; cached boots are ~150ms. Returns the live instance.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequest" }
      responses:
        "201":
          description: Sandbox created and running
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Instance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "400":
          description: Invalid parameters (out-of-range cpu/memory/ports, bad body)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "507":
          description: Host at capacity (MSBD_MAX_SANDBOXES reached). Boot/validation failures are 400/500, not 507.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
    get:
      tags: [lifecycle]
      summary: List all sandboxes
      responses:
        "200":
          description: All known sandboxes (running or stopped)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Instance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/sandboxes/{id}:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
    get:
      tags: [lifecycle]
      summary: Get one sandbox
      responses:
        "200":
          description: Current state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Instance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [lifecycle]
      summary: Delete (tear down) a sandbox
      description: Stops the box if running, then removes it permanently.
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A token from `MSBD_API_KEY` or from the persisted key store
        (`msbd keys create`). Stored tokens are shown once at creation and kept
        only as a SHA-256 hash.
  parameters:
    SandboxId:
      name: id
      in: path
      required: true
      schema: { type: string, example: sbx_1ea598fdaabd2a46 }
      description: Server-generated sandbox id (== microsandbox VM name).
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    NotFound:
      description: Sandbox not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
...
    PayloadTooLarge:
      description: Request body exceeds the configured size limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
...
  schemas:
    Resources:
      type: object
      description: Sizing knobs. Zero/omitted means provider default.
      properties:
        cpu: { type: number, description: vCPUs }
        memory_mb: { type: integer, description: Memory in MiB }
        disk_gb: { type: integer, description: Writable overlay upper size in GiB (OCI images). Zero uses the image default. }
    CreateRequest:
      type: object
      properties:
        image:
          type: string
          description: OCI image reference. Empty → server default (MSBD_DEFAULT_IMAGE).
          example: microsandbox/python
        resources: { $ref: "#/components/schemas/Resources" }
        auto_stop_secs:
          type: integer
          description: Idle seconds before auto-pause. 0 → provider default.
        env:
          type: object
          additionalProperties: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
        workdir:
          type: string
          description: Pin the default cwd. Empty → the image's own WORKDIR is resolved.
        user:
          type: string
          description: Guest user to run as.
        hostname:
          type: string
          description: Guest hostname.
        network_policy:
          type: string
          enum: [none, public-only, allow-all, non-local]
          description: Network egress preset. Empty → provider default.
        ports:
          type: array
          items: { $ref: "#/components/schemas/PortMapping" }
        secrets:
          type: array
          items: { $ref: "#/components/schemas/Secret" }
        mounts:
          type: array
          items: { $ref: "#/components/schemas/Mount" }
    PortMapping:
      type: object
      properties:
        host_port: { type: integer }
        guest_port: { type: integer }
        protocol: { type: string, enum: [tcp, udp] }
    Secret:
      type: object
      properties:
        env_var: { type: string }
        value: { type: string }
    Mount:
      type: object
      properties:
        guest_path: { type: string }
        volume: { type: string, description: Named volume to mount. }
        readonly: { type: boolean }
    Instance:
      type: object
      required: [id, state, workdir]
      properties:
        id: { type: string, example: sbx_1ea598fdaabd2a46 }
        image: { type: string }
        state:
          type: string
          description: Raw microsandbox runtime status, passed through verbatim.
          enum: [running, stopped, crashed, draining, paused, unknown]
        workdir: { type: string, example: /workspace }
        uptime_seconds: { type: number }
        labels:
          type: object
          additionalProperties: { type: string }
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: >
                Machine-readable, cause-oriented error code. Common values:
                bad_request, invalid_params, payload_too_large, unauthorized,
                not_found, forbidden, capacity, create_failed, poll_failed,
                pull_failed, panic, error.
              enum:
                - bad_request
                - invalid_params
                - payload_too_large
                - unauthorized
                - not_found
                - forbidden
                - capacity
                - create_failed
                - list_failed
                - poll_failed
                - metrics_failed
                - remove_failed
                - pull_failed
                - prune_failed
                - reindex_failed
                - not_ready
                - panic
                - error
              example: not_found
            message: { type: string }